python/requirements.txt
selenium==4.22.0
selenium-stealth==1.0.6
python-dotenv==1.0.1
2captcha-python==1.3.0
